+++ src/converters/poetry/dependencies.rs
@@ -4,6 +4,7 @@ use crate::errors::{add_recoverable_error, add_unrecoverable_error};
 use crate::schema;
 use crate::schema::poetry::DependencySpecification;
 use crate::schema::pyproject::DependencyGroupSpecification;
+use crate::schema::utils::SingleOrVec;
 use crate::schema::uv::{SourceContainer, SourceIndex};
 use crate::utils::normalize_dependency_name;
 use indexmap::IndexMap;
@@ -169,6 +170,7 @@ pub fn get_dependency_groups_and_default_groups(
     let mut dependency_groups: IndexMap<String, Vec<DependencyGroupSpecification>> =
         IndexMap::new();
     let mut default_groups: Vec<String> = Vec::new();
+    let mut all_default_groups = false;
 
     // Add dependencies from legacy `[poetry.dev-dependencies]` into `dev` dependency group.
     if let Some(dev_dependencies) = &poetry.dev_dependencies {
@@ -211,6 +213,12 @@ pub fn get_dependency_groups_and_default_groups(
         }
 
         match dependency_groups_strategy {
+            // When using `SetDefaultGroupsAll` strategy, set `default-groups` to "all" under
+            // `[tool.uv]`, to closely match what Poetry does by default, since it includes all
+            // dependency groups.
+            DependencyGroupsStrategy::SetDefaultGroupsAll => {
+                all_default_groups = true;
+            }
             // When using `SetDefaultGroups` strategy, all non-optional dependency groups are
             // referenced in `default-groups` under `[tool.uv]` section. If we only have `dev`
             // dependency group, do not set `default-groups`, as this is already uv's default.
@@ -247,12 +255,13 @@ pub fn get_dependency_groups_and_default_groups(
         return (None, None);
     }
 
-    (
-        Some(dependency_groups),
-        if default_groups.is_empty() {
-            None
-        } else {
-            Some(default_groups)
-        },
-    )
+    let default_groups = if all_default_groups {
+        Some(SingleOrVec::Single("all".to_string()))
+    } else if default_groups.is_empty() {
+        None
+    } else {
+        Some(SingleOrVec::Vec(default_groups))
+    };
+
+    (Some(dependency_groups), default_groups)
 }