+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+* Add `set-default-groups-all` option to `--dependency-groups-strategy` ([#708](https://github.com/mkniewallner/migrate-to-uv/pull/708))
 * Add `--ignore-errors` flag to perform the migration even in case of errors ([#657](https://github.com/mkniewallner/migrate-to-uv/pull/657))
 * [poetry] Bump default `uv_build` bounds to `>=0.10.0,<0.11.0` ([#683](https://github.com/mkniewallner/migrate-to-uv/pull/683))